:lipstick: 채팅 말풍선 컴포넌트 송/수신 분리 및 avatar 추가
