+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@heroui/avatar": "^2.2.12",
     "@heroui/button": "^2.2.13",
     "@heroui/card": "^2.2.15",
     "@heroui/checkbox": "^2.3.9",
